backend_drf_celery/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==4.6.0
psycopg2-binary==2.9.9 